+++ app/models/player.py
@@ -1,0 +1,4 @@
+class Player:
+    def __init__(self, name, finalscore):
+        self.name = name
+        self.finalscore = finalscore
